requirements.txt
mistralai>=0.0.7
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
docling>=0.1.0
openai>=1.0.0
sentence-transformers>=4.1.0
cerebras_cloud_sdk>=1.29.0 